feat: Created initial scorecard_spec tests
Added four tests to the scorecard spec based on user stories 1, 2 and 3. Tests 1,2 and 3 check that the scoredcard logs a score, knows what roll number it is on and also knows what frame number it is on. These tests have passed. Test 4 checks that the scorecard understands that there can only be a maximum of two rolls per frame. Code has not yest been written to pass this test